requirements.txt
flask==3.1.0
openai==1.30.1
httpx==0.27.0
python-dotenv==1.1.0
opencv-python==4.9.0.80
pillow>=10.4.0
numpy==1.26.4
requests==2.31.0
fpdf2
flask-sqlalchemy
pymysql
flask-mail
itsdangerous
gunicorn==21.2.0